join leave into single cog
